Revert change of GetHash impl. signature

+++ src/vmobjects/AbstractObject.cpp
@@ -21,20 +21,20 @@ ComputeSlotMaps(uintptr_t* const root, const vector<uintptr_t*>& fields)
 {
   uintptr_t base = 0;
   uintptr_t slotMap = 0;
-  
-  std::vector<uintptr_t> slotMaps;  
+
+  std::vector<uintptr_t> slotMaps;
 
   for(auto& field : fields) {
     // 1 byte or 8 bytes? assume 1 byte for now, change if incorrect.
     uintptr_t ptrdiff = field - root - base;
-    
+
     if(ptrdiff >= sizeof(uintptr_t) * 8) {
       slotMaps.push_back(slotMap);
       slotMap = 0;
-      
+
       while(ptrdiff >= sizeof(uintptr_t) * 8) {
 	slotMaps.push_back(slotMap);
-	
+
 	base += sizeof(uintptr_t) * 8;
 	ptrdiff -= sizeof(uintptr_t) * 8;
       }
@@ -44,12 +44,12 @@ ComputeSlotMaps(uintptr_t* const root, const vector<uintptr_t*>& fields)
   }
 
   slotMaps.push_back(slotMap);
-  std::reverse(slotMaps.begin(), slotMaps.end());  
+  std::reverse(slotMaps.begin(), slotMaps.end());
   return slotMaps;
 }
 
-size_t AbstractVMObject::GetHash() {
-    return (size_t) this;
+int64_t AbstractVMObject::GetHash() {
+    return (int64_t) this;
 }
 
 void AbstractVMObject::Send(Interpreter* interp, StdString selectorString, vm_oop_t* arguments, long argc) {
